+++ 2024-frontend/src/views/GestionnaireCantineTeledeclarationEnCours.vue
@@ -1,32 +1,22 @@
 <script setup>
-import { computed, onMounted, onUnmounted } from "vue"
+import { computed } from "vue"
 import { storeToRefs } from "pinia"
 import { useRouter } from "vue-router"
 import { useStoreCanteen } from "@/stores/canteen.js"
-import { useStorePurchaseSummary } from "@/stores/purchaseSummary.js"
-import { formatNumber } from "@/utils.js"
 import documentation from "@/data/documentation.json"
 import CanteenSidebarTitle from "@/components/CanteenSidebarTitle.vue"
 import AppHelpCard from "@/components/AppHelpCard.vue"
 import AppBlueCard from "@/components/AppBlueCard.vue"
+import DiagnosticPurchasesLinked from "@/components/DiagnosticPurchasesLinked.vue"
 
 const canteenStore = useStoreCanteen()
 const router = useRouter()
 const currentYear = new Date().getFullYear()
-const lastYear = currentYear - 1
 const { canteenInformations } = storeToRefs(canteenStore)
-
-/* Store */
-const purchaseSummaryStore = useStorePurchaseSummary()
-const { purchaseSummary } = storeToRefs(purchaseSummaryStore)
-onMounted(() => purchaseSummaryStore.initStore(canteenInformations.value.id, lastYear))
-onUnmounted(() => purchaseSummaryStore.deleteStore())
 
 /* Content */
 const pageTitle = computed(() => canteenInformations.value.isGroupe ? `Télédéclaration ${currentYear}` : `Ma télédéclaration ${currentYear}`)
 const firstBlocTitle = computed(() => canteenInformations.value.isGroupe ? 'Bien préparer sa télédéclaration groupée' : 'Bien préparer sa télédéclaration')
-const hasPurchase = computed(() => purchaseSummaryStore.hasPurchaseTotal(lastYear))
-const purchaseAmount = computed(() => `${formatNumber(purchaseSummary.value[lastYear]?.valeurTotale)} €`)
 const hasSatellite = computed(() => canteenInformations.value.isGroupe)
 const satellitesCount = computed(() => canteenInformations.value.satellitesCount)
 const emptySatellitesCount = computed(() => hasSatellite.value && satellitesCount.value == 0)
@@ -104,23 +94,7 @@ const gotToAppro = () => {
       </p>
     </AppBlueCard>
 
-    <AppBlueCard
-      v-if="hasPurchase"
-      class="fr-mt-4w"
-      title="Souhaitez-vous pré-remplir votre déclaration à partir de votre suivi d’achats ?"
-      :alert="{
-        description: 'Optionnel : cette étape n’est pas obligatoire pour déclarer vos approvisionnements.',
-      }"
-      :button="{
-        label: 'Mettre à jour mes achats',
-        to: 'PurchasesHome',
-      }"
-    >
-      <p>
-        Vous avez <strong>{{ purchaseAmount }}</strong> d’achats détectés dans votre suivi des achats. <br />
-        Si vous utilisez l’Outil de Suivi des Achats (Mes achats), pour pré-remplir votre télédéclaration, assurez-vous d’avoir complété l’ensemble de vos achats de l’année précédente.
-      </p>
-    </AppBlueCard>
+    <DiagnosticPurchasesLinked class="fr-mt-4w" :canteen-id="canteenInformations.id" />
 
   </div>
 </template>
